# src/main/java/br/com/ada/f1rst/mail/api/ApiApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.42).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package br.com.ada.f1rst.mail.api;

import br.com.ada.f1rst.mail.model.Error;
import br.com.ada.f1rst.mail.model.*;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import javax.validation.Valid;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2023-04-12T22:46:14.551837857Z[GMT]")
@Validated
public interface ApiApi {

    @Operation(summary = "", description = "", tags={ "Email Management" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Count.class))),
        
        @ApiResponse(responseCode = "204", description = "No Content", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))),
        
        @ApiResponse(responseCode = "500", description = "Internal Server Error", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))) })
    @RequestMapping(value = "/api/email/count",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<?> apiEmailCountGet(@Parameter(in = ParameterIn.QUERY, description = "Sender Address to fetch" ,schema=@Schema()) @Valid @RequestParam(value = "sender_address", required = false) String senderAddress, @Parameter(in = ParameterIn.QUERY, description = "Receiver Address to fetch" ,schema=@Schema()) @Valid @RequestParam(value = "receiver_address", required = false) String receiverAddress, @Parameter(in = ParameterIn.QUERY, description = "Subject contains to fetch" ,schema=@Schema()) @Valid @RequestParam(value = "subject_contains", required = false) String subjectContains, @Parameter(in = ParameterIn.QUERY, description = "Received on the date" ,schema=@Schema()) @Valid @RequestParam(value = "received_date", required = false) String receivedDate);


    @Operation(summary = "", description = "", tags={ "Email Management" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "204", description = "No Content", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))),
        
        @ApiResponse(responseCode = "404", description = "Not found", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))),
        
        @ApiResponse(responseCode = "500", description = "Internal Server Error", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))) })
    @RequestMapping(value = "/api/email",
        produces = { "application/json" }, 
        method = RequestMethod.DELETE)
    ResponseEntity<?> apiEmailDelete(@Parameter(in = ParameterIn.QUERY, description = "Sender Address to fetch" ,schema=@Schema()) @Valid @RequestParam(value = "sender_address", required = false) String senderAddress, @Parameter(in = ParameterIn.QUERY, description = "Receiver Address to fetch" ,schema=@Schema()) @Valid @RequestParam(value = "receiver_address", required = false) String receiverAddress, @Parameter(in = ParameterIn.QUERY, description = "Subject contains to fetch" ,schema=@Schema()) @Valid @RequestParam(value = "subject_contains", required = false) String subjectContains, @Parameter(in = ParameterIn.QUERY, description = "Received on the date" ,schema=@Schema()) @Valid @RequestParam(value = "received_date", required = false) String receivedDate);


    @Operation(summary = "", description = "", tags={ "Email Management" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Emails.class))),
        
        @ApiResponse(responseCode = "204", description = "No Content", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))),
        
        @ApiResponse(responseCode = "500", description = "Internal Server Error", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))) })
    @RequestMapping(value = "/api/email",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<?> apiEmailGet(@Parameter(in = ParameterIn.QUERY, description = "Sender Address to fetch" ,schema=@Schema()) @Valid @RequestParam(value = "sender_address", required = false) String senderAddress, @Parameter(in = ParameterIn.QUERY, description = "Receiver Address to fetch" ,schema=@Schema()) @Valid @RequestParam(value = "receiver_address", required = false) String receiverAddress, @Parameter(in = ParameterIn.QUERY, description = "Subject contains to fetch" ,schema=@Schema()) @Valid @RequestParam(value = "subject_contains", required = false) String subjectContains, @Parameter(in = ParameterIn.QUERY, description = "Received on the date" ,schema=@Schema()) @Valid @RequestParam(value = "received_date", required = false) String receivedDate);


    @Operation(summary = "", description = "", tags={ "Email Management" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Email.class))),
        
        @ApiResponse(responseCode = "404", description = "Not found", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))),
        
        @ApiResponse(responseCode = "500", description = "Internal Server Error", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))) })
    @RequestMapping(value = "/api/email/{id_email}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Email> apiEmailIdEmailGet(@Parameter(in = ParameterIn.PATH, description = "Email identifier", required=true, schema=@Schema()) @PathVariable("id_email") Long idEmail);


    @Operation(summary = "", description = "", tags={ "Email Management" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Created.class))),
        
        @ApiResponse(responseCode = "400", description = "Bad Request", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))),
        
        @ApiResponse(responseCode = "500", description = "Internal Server Error", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))) })
    @RequestMapping(value = "/api/email",
        produces = { "application/json" }, 
        consumes = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<Created> apiEmailPost(@Parameter(in = ParameterIn.DEFAULT, description = "", schema=@Schema()) @Valid @RequestBody Email body);

}