# src/main/java/br/com/ada/f1rst/mail/model/Count.java
package br.com.ada.f1rst.mail.model;

import lombok.*;
import org.springframework.validation.annotation.Validated;

@Validated
@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@ToString
@EqualsAndHashCode
@Builder
public class Count   {

  private Integer quantity;
}