# src/main/java/br/com/ada/f1rst/mail/model/Created.java
package br.com.ada.f1rst.mail.model;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.*;
import org.springframework.validation.annotation.Validated;

@Validated
@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@ToString
@EqualsAndHashCode
@Builder
public class Created   {

  @JsonProperty("new_id")
  private Long newId;
}